Fixed layout of page to be more intuitive
