+++ dashboard/app.py
@@ -4,203 +4,27 @@ A multi-page dashboard for tracking social media trends and analytics.
 Sources of data are bluesky and Google trends.
 """
 
-import hashlib
-import hmac
 import logging
-import re
-import secrets
-from datetime import datetime, timedelta
-from os import environ as ENV, _Environ
-from typing import Optional
-import random
-
-import altair as alt
-import pandas as pd
-import psycopg2
-from psycopg2.extras import RealDictCursor
 import streamlit as st
-from dotenv import load_dotenv
+from psycopg2.extras import RealDictCursor
+
+# Import shared utilities
+from utils import (
+    get_db_connection,
+    get_db_connection_cleanup,
+    get_user_by_username,
+    authenticate_user,
+    generate_password_hash,
+    validate_signup_input,
+    create_user
+)
 
 # ============== Logging Configuration ==============
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-# ============== Environment Configuration ==============
-load_dotenv()
-
-# ============== Database Configuration ==============
-DB_CONFIG = {
-    "host": ENV.get("DB_HOST"),
-    "port": int(ENV.get("DB_PORT", 5432)),
-    "database": ENV.get("DB_NAME"),
-    "user": ENV.get("DB_USER"),
-    "password": ENV.get("DB_PASSWORD")
-}
-
-@st.cache_resource
-def get_db_connection() -> Optional[psycopg2.extensions.connection]:
-    """Get persistent database connection for the entire session."""
-    conn = psycopg2.connect(
-        host=DB_CONFIG.get("host"),
-        port=DB_CONFIG.get("port", 5432),
-        database=DB_CONFIG.get("database"),
-        user=DB_CONFIG.get("user"),
-        password=DB_CONFIG.get("password")
-    )
-    logger.info("Database connection established.")
-    return conn
-
-
-def get_user_by_username(cursor, email: str) -> Optional[dict]:
-    """Retrieve user details from database by username."""
-    query = "SELECT * FROM users WHERE email = %s"
-    cursor.execute(query, (email,))
-    result = cursor.fetchone()
-    return result
-
-
-def verify_password(stored_hash: str, entered_password: str) -> bool:
-    """Verify if entered password matches the stored hash. Uses PBKDF2-SHA256 for password hashing."""
-    parts = stored_hash.split("$")
-    if len(parts) != 3:
-        return False
-
-    salt, iterations_str, stored_hash_hex = parts
-
-    # Validate iterations is a valid integer
-    if not iterations_str.isdigit():
-        return False
-
-    iterations = int(iterations_str)
-
-    # Hash the entered password with the same salt
-    hashed = hashlib.pbkdf2_hmac(
-        "sha256",
-        entered_password.encode("utf-8"),
-        salt.encode("utf-8"),
-        iterations
-    )
-
-    # Compare hashes using time constant comparison to prevent timing attacks
-    return hmac.compare_digest(hashed.hex(), stored_hash_hex)
-
-
-def authenticate_user(cursor, username: str, password: str) -> bool:
-    """Authenticate user by checking username and password."""
-    user = get_user_by_username(cursor, username)
-
-    if user is None:
-        return False
-
-    return verify_password(user["password_hash"], password)
-
-
-def generate_password_hash(password: str, iterations: int = 100000) -> str:
-    """Generate a password hash using PBKDF2-SHA256."""
-    if not password:
-        raise ValueError("Password cannot be empty")
-
-    # Generate a random salt
-    salt = secrets.token_hex(16)
-
-    # Hash the password
-    hashed = hashlib.pbkdf2_hmac(
-        "sha256",
-        password.encode("utf-8"),
-        salt.encode("utf-8"),
-        iterations
-    )
-
-    # Return formatted hash
-    return f"{salt}${iterations}${hashed.hex()}"
-
-
-def validate_signup_input(email: str, password: str) -> bool:
-    """Validate signup input: email format and password length."""
-    if not email or not password:
-        return False
-
-    # Validate email format (basic check)
-    email_pattern = r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$"
-    if not re.match(email_pattern, email):
-        return False
-
-    # Validate password length
-    if len(password) <= 8:
-        return False
-
-    return True
-
-
-def create_user(cursor, email: str, password_hash: str) -> bool:
-    """Insert a new user into the database."""
-    try:
-        query = "INSERT INTO users (email, password_hash) VALUES (%s, %s)"
-        cursor.execute(query, (email, password_hash))
-        cursor.connection.commit()
-        return True
-    except psycopg2.IntegrityError:
-        # Email already exists
-        cursor.connection.rollback()
-        return False
-    except psycopg2.Error as e:
-        logger.error(f"Database error creating user: {e}")
-        cursor.connection.rollback()
-        return False
-
-
-@st.cache_resource
-def get_db_connection_cleanup():
-    """Register cleanup for database connection on app exit."""
-    def close_conn():
-        conn = get_db_connection()
-        if conn:
-            conn.close()
-            logger.info("Database connection closed.")
-    return close_conn
-
-
 # Register cleanup
 _cleanup = get_db_connection_cleanup()
-
-
-# ============== Keyword Management Functions ==============
-
-def add_user_keyword(cursor, user_id: int, keyword: str) -> bool:
-    """Add a keyword to a user's tracked keywords."""
-    # Insert keyword into keywords table (case-insensitive, do nothing on conflict)
-    cursor.execute(
-        "INSERT INTO keywords (keyword_value) VALUES (LOWER(%s)) ON CONFLICT (keyword_value) DO NOTHING",
-        (keyword,)
-    )
-
-    # Add entry to user_keywords table mapping user to keyword
-    cursor.execute(
-        "INSERT INTO user_keywords (user_id, keyword_id) SELECT %s, keyword_id FROM keywords WHERE LOWER(keyword_value) = LOWER(%s)",
-        (user_id, keyword)
-    )
-    cursor.connection.commit()
-    return True
-
-
-def remove_user_keyword(cursor, user_id: int, keyword: str) -> bool:
-    """Remove a keyword from a user's tracked keywords."""
-    cursor.execute(
-        "DELETE FROM user_keywords WHERE user_id = %s AND keyword_id IN (SELECT keyword_id FROM keywords WHERE LOWER(keyword_value) = LOWER(%s))",
-        (user_id, keyword)
-    )
-    cursor.connection.commit()
-    return True
-
-
-def get_user_keywords(cursor, user_id: int) -> list:
-    """Retrieve all keywords for a user."""
-    cursor.execute(
-        "SELECT k.keyword_value FROM keywords k JOIN user_keywords uk ON k.keyword_id = uk.keyword_id WHERE uk.user_id = %s ORDER BY k.keyword_value",
-        (user_id,)
-    )
-    results = cursor.fetchall()
-    return [row["keyword_value"] for row in results] if results else []
 
 
 # ============== Page Configuration ==============
@@ -208,7 +32,7 @@ st.set_page_config(
     page_title="Trends Tracker",
     page_icon="📊",
     layout="wide",
-    initial_sidebar_state="expanded"
+    initial_sidebar_state="collapsed"
 )
 
 # ============== Session State Initialization ==============
@@ -218,8 +42,6 @@ if "username" not in st.session_state:
     st.session_state.username = ""
 if "user_id" not in st.session_state:
     st.session_state.user_id = None
-if "keywords" not in st.session_state:
-    st.session_state.keywords = ["matcha", "boba", "coffee"]
 if "emails_enabled" not in st.session_state:
     st.session_state.emails_enabled = False
 if "alerts_enabled" not in st.session_state:
@@ -228,258 +50,18 @@ if "email" not in st.session_state or not st.session_state.logged_in:
     st.session_state.email = ""
 
 
-# ============== Placeholder Data Generators ==============
-def generate_placeholder_metrics(keyword: str, days: int) -> dict:
-    """Generate placeholder KPI metrics for a keyword."""
-    random.seed(hash(keyword) + days)
-    return {
-        "mentions": random.randint(500, 5000),
-        "posts": random.randint(200, 2000),
-        "reposts": random.randint(100, 1500),
-        "comments": random.randint(150, 1800),
-        "avg_sentiment": round(random.uniform(-0.3, 0.7), 2)
-    }
-
-
-def generate_time_series_data(keyword: str, days: int) -> pd.DataFrame:
-    """Generate placeholder time series data for activity metrics."""
-    random.seed(hash(keyword))
-    dates = [datetime.now() - timedelta(days=x) for x in range(days, 0, -1)]
-    data = {
-        "date": dates,
-        "posts": [random.randint(10, 100) for _ in dates],
-        "reposts": [random.randint(5, 80) for _ in dates],
-        "comments": [random.randint(8, 90) for _ in dates],
-    }
-    return pd.DataFrame(data)
-
-
-def generate_sentiment_breakdown(keyword: str) -> dict:
-    """Generate placeholder sentiment distribution."""
-    random.seed(hash(keyword))
-    positive = random.randint(30, 50)
-    negative = random.randint(10, 30)
-    neutral = 100 - positive - negative
-    return {"Positive": positive, "Neutral": neutral, "Negative": negative}
-
-
-def generate_keywords_summary(keywords: list, days: int) -> pd.DataFrame:
-    """Generate summary table for all keywords."""
-    data = []
-    for kw in keywords:
-        metrics = generate_placeholder_metrics(kw, days)
-        data.append({
-            "Keyword": kw,
-            "Mentions": metrics["mentions"],
-            "Posts": metrics["posts"],
-            "Reposts": metrics["reposts"],
-            "Comments": metrics["comments"],
-            "Avg Sentiment": metrics["avg_sentiment"]
-        })
-    return pd.DataFrame(data)
-
-
-def generate_llm_summary(keyword: str, days: int) -> str:
-    """Generate placeholder LLM summary."""
-    return f"""
-    **📝 AI Summary for "{keyword}" (Last {days} days)**
-
-    The topic "{keyword}" has shown moderate engagement over the selected period.
-    Overall sentiment appears to be leaning positive with a notable increase in mentions
-    during peak hours. Key themes include product reviews, lifestyle content, and
-    community discussions. Consider monitoring related trending terms for broader insights.
-
-    *[This is a placeholder summary - will be generated by LLM integration]*
-    """
-
-
-# ============== Visualization Data Generators ==============
-
-def generate_word_cloud_data(keyword: str, days: int) -> dict:
-    """Generate word frequency data for word cloud visualization."""
-    random.seed(hash(keyword) + days)
-
-    # Associated words vary by keyword
-    word_pools = {
-        "matcha": ["latte", "green", "tea", "healthy", "cafe", "japanese", "powder", "organic", "iced", "oat"],
-        "boba": ["milk", "tea", "tapioca", "pearls", "sweet", "bubble", "drink", "asian", "chewy", "sugar"],
-        "coffee": ["espresso", "latte", "beans", "morning", "caffeine", "roast", "brew", "americano", "mocha", "cold"],
-    }
-
-    # Use keyword-specific words or generic ones
-    base_words = word_pools.get(keyword.lower(), [
-        "trending", "popular", "viral", "discussed", "mentioned", "shared", "liked", "commented", "posted", "engaged"
-    ])
-
-    # Add some common words
-    common_words = ["love", "great", "best", "new", "try", "amazing", "delicious", "perfect", "favorite", "recommend"]
-    all_words = base_words + common_words
-
-    # Generate frequencies
-    return {word: random.randint(10, 100) for word in all_words}
-
-
-def generate_sentiment_calendar_data(keyword: str, days: int) -> pd.DataFrame:
-    """Generate daily sentiment data for calendar heatmap visualization."""
-    random.seed(hash(keyword) + days)
-
-    dates = [datetime.now() - timedelta(days=x) for x in range(days, 0, -1)]
-
-    data = []
-    for date in dates:
-        sentiment = round(random.uniform(-0.5, 0.8), 2)
-        data.append({
-            "date": date,
-            "sentiment": sentiment,
-            "day_of_week": date.weekday(),
-            "week": date.isocalendar()[1]
-        })
-
-    return pd.DataFrame(data)
-
-
-def generate_trending_velocity(keyword: str, days: int) -> dict:
-    """Generate trending velocity data for speedometer visualization."""
-    random.seed(hash(keyword) + days)
-
-    velocity = random.randint(20, 85)
-    percent_change = round(random.uniform(-30, 50), 1)
-
-    if percent_change > 10:
-        direction = "accelerating"
-    elif percent_change < -10:
-        direction = "decelerating"
-    else:
-        direction = "stable"
-
-    return {
-        "velocity": velocity,
-        "direction": direction,
-        "percent_change": percent_change,
-        "current_mentions": random.randint(500, 3000),
-        "previous_mentions": random.randint(400, 2500)
-    }
-
-
-def generate_network_graph_data(keywords: list) -> dict:
-    """Generate node and edge data for keyword network graph visualization."""
-    if not keywords:
-        return {"nodes": [], "edges": []}
-
-    random.seed(hash(tuple(keywords)))
-
-    # Create nodes for each keyword
-    nodes = [{"id": kw, "label": kw, "size": random.randint(20, 50)} for kw in keywords]
-
-    # Add related topic nodes
-    related_topics = {
-        "matcha": ["latte", "green tea", "healthy"],
-        "boba": ["milk tea", "tapioca", "asian drinks"],
-        "coffee": ["espresso", "caffeine", "morning"]
-    }
-
-    for kw in keywords:
-        topics = related_topics.get(kw.lower(), ["topic1", "topic2"])
-        for topic in topics[:2]:  # Limit to 2 related topics per keyword
-            nodes.append({"id": topic, "label": topic, "size": random.randint(10, 30)})
-
-    # Create edges between keywords and their related topics
-    edges = []
-    for kw in keywords:
-        topics = related_topics.get(kw.lower(), ["topic1", "topic2"])
-        for topic in topics[:2]:
-            edges.append({
-                "source": kw,
-                "target": topic,
-                "weight": random.randint(5, 30)
-            })
-
-    # Add edges between keywords that co-occur
-    for i, kw1 in enumerate(keywords):
-        for kw2 in keywords[i+1:]:
-            if random.random() > 0.5:  # 50% chance of connection
-                edges.append({
-                    "source": kw1,
-                    "target": kw2,
-                    "weight": random.randint(10, 40)
-                })
-
-    return {"nodes": nodes, "edges": edges}
-
-
-def generate_random_post(keyword: str) -> dict:
-    """Generate a random post for typing animation display."""
-    random.seed(hash(keyword) + int(datetime.now().timestamp() // 60))  # Changes every minute
-
-    post_templates = [
-        f"Just discovered the best {keyword} spot in town! Highly recommend 🙌",
-        f"Anyone else obsessed with {keyword} lately? Can't get enough!",
-        f"Hot take: {keyword} is overrated. Change my mind 🤔",
-        f"My morning routine isn't complete without {keyword} ☀️",
-        f"Tried a new {keyword} recipe today and it was amazing!",
-        f"The {keyword} trend is everywhere and I'm here for it 💯",
-        f"Unpopular opinion: {keyword} hits different at night",
-        f"Finally found a place that does {keyword} right! 🎯",
-    ]
-
-    authors = ["@trendwatcher", "@foodie_life", "@daily_vibes", "@lifestyle_guru", "@taste_explorer"]
-
-    return {
-        "text": random.choice(post_templates),
-        "author": random.choice(authors),
-        "timestamp": datetime.now() - timedelta(hours=random.randint(1, 48)),
-        "likes": random.randint(50, 500),
-        "reposts": random.randint(10, 100)
-    }
-
-
-def generate_ai_insights(keyword: str, days: int) -> dict:
-    """Generate AI-powered insights for a keyword."""
-    random.seed(hash(keyword) + days)
-
-    themes = [
-        "Product reviews and recommendations",
-        "Lifestyle and wellness content",
-        "Community discussions",
-        "Recipe sharing",
-        "Price comparisons"
-    ]
-
-    positive_drivers = [
-        "High-quality product mentions",
-        "Positive customer experiences",
-        "Influencer endorsements"
-    ]
-
-    negative_drivers = [
-        "Price concerns",
-        "Availability issues",
-        "Quality inconsistencies"
-    ]
-
-    recommendations = [
-        f"Monitor '{keyword} alternatives' - emerging search term",
-        f"Peak engagement for {keyword} occurs between 9-11 AM",
-        f"Consider tracking related hashtags for broader reach",
-        f"Sentiment dips on weekends - investigate weekend-specific concerns"
-    ]
-
-    summary = f"""The topic "{keyword}" has shown {'strong' if random.random() > 0.5 else 'moderate'} engagement over the last {days} days. Overall sentiment is {'predominantly positive' if random.random() > 0.4 else 'mixed'}, with notable spikes during {'morning' if random.random() > 0.5 else 'evening'} hours. Key discussions center around product quality and lifestyle integration. {'An emerging trend shows increased interest in premium variants.' if random.random() > 0.5 else 'Community-driven content is gaining traction.'}"""
-
-    return {
-        "summary": summary,
-        "themes": random.sample(themes, k=3),
-        "sentiment_drivers": {
-            "positive": random.sample(positive_drivers, k=2),
-            "negative": random.sample(negative_drivers, k=2)
-        },
-        "recommendations": random.sample(recommendations, k=3)
-    }
-
-
 # ============== Login Page ==============
 def show_login_page():
     """Display the login/signup page."""
+    # Hide sidebar on login page
+    st.markdown("""
+        <style>
+            [data-testid="stSidebar"] {
+                display: none;
+            }
+        </style>
+    """, unsafe_allow_html=True)
+
     st.title("🔐 Trends Tracker")
     st.markdown("### Welcome! Please login or create an account to continue.")
 
@@ -508,6 +90,10 @@ def show_login_page():
                             st.session_state.username = login_username
                             st.session_state.user_id = user["user_id"]
                             st.session_state.email = user["email"]
+                            # Clear user-specific data from previous sessions
+                            st.session_state.keywords = []
+                            st.session_state.keywords_loaded = False
+                            st.session_state.alerts_loaded = False
                             st.success("Login successful!")
                             st.rerun()
                         else:
@@ -551,6 +137,10 @@ def show_login_page():
                             st.session_state.username = signup_name.split()[0]
                             st.session_state.user_id = user["user_id"]
                             st.session_state.email = user["email"]
+                            # Clear user-specific data from previous sessions
+                            st.session_state.keywords = []
+                            st.session_state.keywords_loaded = False
+                            st.session_state.alerts_loaded = False
                             st.success("Account created successfully!")
                             st.rerun()
                         else:
@@ -560,356 +150,12 @@ def show_login_page():
                         st.error("Unable to connect to database. Please try again later.")
 
 
-# ============== Main Analytics Dashboard ==============
-def show_analytics_dashboard():
-    """Display the main analytics dashboard."""
-    st.title("📊 Analytics Dashboard")
-
-    # Top controls row
-    col1, col2, col3 = st.columns([2, 2, 4])
-    with col1:
-        selected_keyword = st.selectbox(
-            "Select Keyword",
-            options=st.session_state.keywords,
-            index=0
-        )
-    with col2:
-        days_options = {"Last 7 days": 7, "Last 14 days": 14, "Last 30 days": 30, "Last 90 days": 90}
-        selected_period = st.selectbox("Time Period", options=list(days_options.keys()))
-        days = days_options[selected_period]
-
-    st.markdown("---")
-
-    # LLM Summary Section
-    st.markdown(generate_llm_summary(selected_keyword, days))
-
-    st.markdown("---")
-
-    # KPI Metrics Row
-    metrics = generate_placeholder_metrics(selected_keyword, days)
-
-    kpi_col1, kpi_col2, kpi_col3, kpi_col4, kpi_col5 = st.columns(5)
-
-    with kpi_col1:
-        st.metric(
-            label="📢 Mentions",
-            value=f"{metrics['mentions']:,}",
-            delta=f"{random.randint(-10, 20)}%"
-        )
-    with kpi_col2:
-        st.metric(
-            label="📝 Posts",
-            value=f"{metrics['posts']:,}",
-            delta=f"{random.randint(-10, 20)}%"
-        )
-    with kpi_col3:
-        st.metric(
-            label="🔄 Reposts",
-            value=f"{metrics['reposts']:,}",
-            delta=f"{random.randint(-10, 20)}%"
-        )
-    with kpi_col4:
-        st.metric(
-            label="💬 Comments",
-            value=f"{metrics['comments']:,}",
-            delta=f"{random.randint(-10, 20)}%"
-        )
-    with kpi_col5:
-        sentiment_color = "normal" if metrics['avg_sentiment'] >= 0 else "inverse"
-        st.metric(
-            label="😊 Avg Sentiment",
-            value=f"{metrics['avg_sentiment']:.2f}",
-            delta=f"{random.uniform(-0.1, 0.1):.2f}",
-            delta_color=sentiment_color
-        )
-
-    st.markdown("---")
-
-    # Charts Row 1: Activity Over Time & Sentiment Breakdown
-    chart_col1, chart_col2 = st.columns(2)
-
-    with chart_col1:
-        st.markdown("### 📈 Activity Metrics Over Time")
-        time_data = generate_time_series_data(selected_keyword, days)
-
-        # Melt data for Altair multi-line chart
-        time_data_melted = time_data.melt(
-            id_vars=["date"],
-            value_vars=["posts", "reposts", "comments"],
-            var_name="Metric",
-            value_name="Count"
-        )
-
-        line_chart = alt.Chart(time_data_melted).mark_line(point=True).encode(
-            x=alt.X("date:T", title="Date"),
-            y=alt.Y("Count:Q", title="Count"),
-            color=alt.Color("Metric:N", scale=alt.Scale(
-                domain=["posts", "reposts", "comments"],
-                range=["#636EFA", "#EF553B", "#00CC96"]
-            )),
-            tooltip=["date:T", "Metric:N", "Count:Q"]
-        ).properties(height=350).interactive()
-
-        st.altair_chart(line_chart, use_container_width=True)
-
-    with chart_col2:
-        st.markdown("### 🎭 Sentiment Breakdown")
-        sentiment_data = generate_sentiment_breakdown(selected_keyword)
-
-        sentiment_df = pd.DataFrame({
-            "Sentiment": list(sentiment_data.keys()),
-            "Percentage": list(sentiment_data.values())
-        })
-
-        pie_chart = alt.Chart(sentiment_df).mark_arc(innerRadius=50).encode(
-            theta=alt.Theta("Percentage:Q"),
-            color=alt.Color("Sentiment:N", scale=alt.Scale(
-                domain=["Positive", "Neutral", "Negative"],
-                range=["#00CC96", "#636EFA", "#EF553B"]
-            )),
-            tooltip=["Sentiment:N", "Percentage:Q"]
-        ).properties(height=350)
-
-        st.altair_chart(pie_chart, use_container_width=True)
-
-    st.markdown("---")
-
-    # Charts Row 2: Overall Sentiment & Daily Averages
-    chart_col3, chart_col4 = st.columns(2)
-
-    with chart_col3:
-        st.markdown("### 📊 Overall Sentiment Score Over Time")
-        time_data["sentiment"] = [round(random.uniform(-0.5, 0.8), 2) for _ in range(len(time_data))]
-
-        # Area chart with sentiment
-        sentiment_area = alt.Chart(time_data).mark_area(
-            line={"color": "#AB63FA"},
-            color=alt.Gradient(
-                gradient="linear",
-                stops=[
-                    alt.GradientStop(color="white", offset=0),
-                    alt.GradientStop(color="#AB63FA", offset=1)
-                ],
-                x1=1, x2=1, y1=1, y2=0
-            )
-        ).encode(
-            x=alt.X("date:T", title="Date"),
-            y=alt.Y("sentiment:Q", title="Sentiment Score", scale=alt.Scale(domain=[-1, 1])),
-            tooltip=["date:T", "sentiment:Q"]
-        ).properties(height=300)
-
-        # Add horizontal line at y=0
-        zero_line = alt.Chart(pd.DataFrame({"y": [0]})).mark_rule(
-            strokeDash=[5, 5], color="gray"
-        ).encode(y="y:Q")
-
-        st.altair_chart(sentiment_area + zero_line, use_container_width=True)
-
-    with chart_col4:
-        st.markdown("### 📅 Daily Average by Metric Type")
-        avg_data = pd.DataFrame({
-            "Metric": ["Posts", "Reposts", "Comments"],
-            "Daily Avg": [
-                round(time_data["posts"].mean(), 1),
-                round(time_data["reposts"].mean(), 1),
-                round(time_data["comments"].mean(), 1)
-            ]
-        })
-
-        bar_chart = alt.Chart(avg_data).mark_bar().encode(
-            x=alt.X("Metric:N", title="Metric"),
-            y=alt.Y("Daily Avg:Q", title="Daily Average"),
-            color=alt.Color("Metric:N", scale=alt.Scale(
-                domain=["Posts", "Reposts", "Comments"],
-                range=["#636EFA", "#EF553B", "#00CC96"]
-            ), legend=None),
-            tooltip=["Metric:N", "Daily Avg:Q"]
-        ).properties(height=300)
-
-        st.altair_chart(bar_chart, use_container_width=True)
-
-    st.markdown("---")
-
-    # Chart Row 3: Keyword Comparison
-    st.markdown("### 🔍 Keyword Mentions Comparison")
-    if len(st.session_state.keywords) > 0:
-        comparison_data = pd.DataFrame({
-            "Keyword": st.session_state.keywords,
-            "Mentions": [generate_placeholder_metrics(kw, days)["mentions"] for kw in st.session_state.keywords]
-        })
-
-        comparison_chart = alt.Chart(comparison_data).mark_bar().encode(
-            x=alt.X("Keyword:N", title="Keyword"),
-            y=alt.Y("Mentions:Q", title="Mentions"),
-            color=alt.Color("Keyword:N", scale=alt.Scale(scheme="set2"), legend=None),
-            tooltip=["Keyword:N", "Mentions:Q"]
-        ).properties(height=300)
-
-        st.altair_chart(comparison_chart, use_container_width=True)
-    else:
-        st.info("Add keywords in the Topics tab to see comparison data.")
-
-    st.markdown("---")
-
-    # Summary Table
-    st.markdown("### 📋 Keywords Summary Table")
-    if len(st.session_state.keywords) > 0:
-        summary_df = generate_keywords_summary(st.session_state.keywords, days)
-
-        # Style the dataframe
-        def color_sentiment(val):
-            if val > 0.2:
-                return "background-color: #d4edda"
-            elif val < -0.2:
-                return "background-color: #f8d7da"
-            return ""
-
-        styled_df = summary_df.style.applymap(
-            color_sentiment, subset=["Avg Sentiment"]
-        ).format({"Avg Sentiment": "{:.2f}"})
-
-        st.dataframe(styled_df, use_container_width=True, hide_index=True)
-    else:
-        st.info("Add keywords in the Topics tab to see summary data.")
-
-
-# ============== Topics Dashboard ==============
-def show_topics_dashboard():
-    """Display the topics/keywords management dashboard."""
-    st.title("🏷️ Topics Management")
-    st.markdown("Manage your tracked keywords and topics here.")
-
-    st.markdown("---")
-
-    # Load keywords from database on first visit
-    if "keywords_loaded" not in st.session_state:
-        conn = get_db_connection()
-        if conn and st.session_state.user_id:
-            cursor = conn.cursor(cursor_factory=RealDictCursor)
-            db_keywords = get_user_keywords(cursor, st.session_state.user_id)
-            cursor.close()
-            st.session_state.keywords = db_keywords
-            st.session_state.keywords_loaded = True
-
-    # Add new keyword section
-    st.markdown("### ➕ Add New Keyword")
-    col1, col2 = st.columns([3, 1])
-
-    with col1:
-        new_keyword = st.text_input(
-            "Enter keyword",
-            placeholder="e.g., matcha, tea, boba...",
-            label_visibility="collapsed"
-        )
-    with col2:
-        if st.button("Add Keyword", type="primary", use_container_width=True):
-            if new_keyword:
-                if new_keyword.lower() not in [k.lower() for k in st.session_state.keywords]:
-                    # Add to database
-                    conn = get_db_connection()
-                    if conn and st.session_state.user_id:
-                        cursor = conn.cursor(cursor_factory=RealDictCursor)
-                        add_user_keyword(cursor, st.session_state.user_id, new_keyword)
-                        cursor.close()
-
-                    # Add to session state
-                    st.session_state.keywords.append(new_keyword.lower())
-                    st.success(f'Added "{new_keyword}" to your keywords!')
-                    st.rerun()
-                else:
-                    st.warning("This keyword already exists.")
-            else:
-                st.error("Please enter a keyword.")
-
-    st.markdown("---")
-
-    # Current keywords display
-    st.markdown("### 📝 Your Keywords")
-
-    if len(st.session_state.keywords) > 0:
-        # Display keywords as a grid with delete buttons
-        cols = st.columns(4)
-        for idx, keyword in enumerate(st.session_state.keywords):
-            with cols[idx % 4]:
-                col_a, col_b = st.columns([3, 1])
-                with col_a:
-                    st.markdown(f"""
-                    <div style="
-                        background-color: #f0f2f6;
-                        padding: 10px 15px;
-                        border-radius: 20px;
-                        margin: 5px 0;
-                        text-align: center;
-                        font-weight: 500;
-                    ">
-                        {keyword}
-                    </div>
-                    """, unsafe_allow_html=True)
-                with col_b:
-                    if st.button("🗑️", key=f"delete_{idx}"):
-                        # Remove from database
-                        conn = get_db_connection()
-                        if conn and st.session_state.user_id:
-                            cursor = conn.cursor(cursor_factory=RealDictCursor)
-                            remove_user_keyword(cursor, st.session_state.user_id, keyword)
-                            cursor.close()
-
-                        # Remove from session state
-                        st.session_state.keywords.remove(keyword)
-                        st.rerun()
-
-        st.markdown("---")
-
-# ============== Sidebar Navigation ==============
-def show_sidebar():
-    """Display the sidebar navigation."""
-    with st.sidebar:
-        st.markdown(f"### 👋 Hello, {st.session_state.username}!")
-        st.markdown("---")
-
-        # Navigation
-        st.markdown("### 🧭 Navigation")
-        page = st.radio(
-            "Go to",
-            options=["📊 Analytics", "🏷️ Topics", "🔔 Alerts"],
-            label_visibility="collapsed"
-        )
-
-        st.markdown("---")
-
-        # Quick stats
-        st.markdown("### 📈 Quick Stats")
-        st.metric("Keywords Tracked", len(st.session_state.keywords))
-        st.metric("Alerts Enabled", "Yes" if st.session_state.alerts_enabled else "No")
-
-        st.markdown("---")
-
-        # Logout button
-        if st.button("🚪 Logout", use_container_width=True):
-            st.session_state.logged_in = False
-            st.session_state.username = ""
-            st.session_state.user_id = None
-            st.rerun()
-
-        st.markdown("---")
-        st.caption("Trends Tracker v1.0 (Mockup)")
-
-    return page
-
-
 # ============== Main App ==============
 def main():
     """Main application entry point."""
     if not st.session_state.logged_in:
         show_login_page()
     else:
-        page = show_sidebar()
-
-        if page == "📊 Analytics":
-            show_analytics_dashboard()
-        elif page == "🏷️ Topics":
-            show_topics_dashboard()
-        # Redirect to Home page after login
         st.switch_page("pages/1_Home.py")
