+++ dashboard/reqs_dashboard.txt
@@ -5,6 +5,9 @@ python-dotenv
 pytest
 streamlit
 
+boto3
+botocore[crt]
+
 # Visualization libraries
 wordcloud
 matplotlib
